+++ app/Transformer/Api/AccountTransformer.php
@@ -31,6 +31,7 @@ class AccountTransformer extends Fractal\TransformerAbstract
 			'url' => $profile->url(),
 			'avatar' => $profile->avatarUrl(),
 			'website' => $profile->website,
+			'local' => (bool) $local,
 			'is_admin' => (bool) $is_admin,
 		];
 	}