+++ resources/lang/de/auth.php
@@ -14,6 +14,6 @@ return [
     */
 
     'failed'   => 'Diese Anmeldeinformationen stimmen nicht mit unseren Daten überein.',
-    'throttle' => 'Zu viele Login-Versuche. Versuche es in :seconds Sekunden erneut.',
+    'throttle' => 'Zu viele Anmeldeversuche. Versuche es in :seconds Sekunden erneut.',
 
 ];