+++ resources/lang/de/site.php
@@ -8,12 +8,12 @@ return [
 	'fediverse'		=> 'Fediverse',
 	'opensource'	=> 'Open Source',
 	'terms'			=> 'Nutzungshinweise',
-	'privacy'		=> 'Privacy',
+	'privacy'		=> 'Datenschutz',
 	'l10nWip'		=> 'Wir arbeiten noch an der Unterstützung weiterer Sprachen',
 	'currentLocale' => 'Aktuelle Sprache',
 	'selectLocale'  => 'Wähle eine der unterstützten Sprachen aus',
 	'contact'		=> 'Kontakt',
 	'contact-us'	=> 'Kontaktiere uns',
-	'places'		=> 'Plätze',
+	'places'		=> 'Orte',
 
 ];