+++ tests/Feature/LoginTest.php
@@ -5,6 +5,7 @@ namespace Tests\Feature;
 use Tests\TestCase;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Illuminate\Foundation\Testing\WithoutMiddleware;
+use App\User;
 
 class LoginTest extends TestCase
 {
@@ -15,19 +16,5 @@ class LoginTest extends TestCase
         $response = $this->get('login');
 
         $response->assertSee('Forgot Password');
-    }
-
-    /** @test */
-    public function view_register_page()
-    {
-        if(true == config('pixelfed.open_registration')) {
-            $response = $this->get('register');
-
-            $response->assertSee('Register a new account');
-        } else {
-            $response = $this->get('register');
-
-            $response->assertSee('Registration is closed');
-        }
     }
 }